requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.0.1
python-decouple==3.8
django-filter==23.3
django-tables2==2.6.0
django-bootstrap5==23.1
